move SceneUpdateCache into its on file
