+++ CMakeLists.txt
@@ -411,6 +411,7 @@ target_sources(${FGE_SERVER_LIB_NAME} PRIVATE "sources/C_task.cpp")
 
 target_sources(${FGE_SERVER_LIB_NAME} PRIVATE "sources/object/C_childObjectsAccessor.cpp")
 target_sources(${FGE_SERVER_LIB_NAME} PRIVATE "sources/C_scene.cpp")
+target_sources(${FGE_SERVER_LIB_NAME} PRIVATE "sources/C_sceneUpdateCache.cpp")
 target_sources(${FGE_SERVER_LIB_NAME} PRIVATE "sources/C_subscription.cpp")
 target_sources(${FGE_SERVER_LIB_NAME} PRIVATE "sources/C_tagList.cpp")
 target_sources(${FGE_SERVER_LIB_NAME} PRIVATE "sources/C_texture.cpp")
@@ -521,6 +522,7 @@ target_sources(${FGE_LIB_NAME} PRIVATE "sources/C_task.cpp")
 
 target_sources(${FGE_LIB_NAME} PRIVATE "sources/object/C_childObjectsAccessor.cpp")
 target_sources(${FGE_LIB_NAME} PRIVATE "sources/C_scene.cpp")
+target_sources(${FGE_LIB_NAME} PRIVATE "sources/C_sceneUpdateCache.cpp")
 target_sources(${FGE_LIB_NAME} PRIVATE "sources/C_soundBuffer.cpp")
 target_sources(${FGE_LIB_NAME} PRIVATE "sources/C_subscription.cpp")
 target_sources(${FGE_LIB_NAME} PRIVATE "sources/C_tagList.cpp")
